algo_app/requirements.txt

boto3
python-dotenv==1.0.0





certifi==2022.12.7

charset-normalizer==3.1.0



dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0




gunicorn==20.1.0

idna==3.4

jsonschema==4.17.3


numpy==1.23.4

pandas==1.5.3


plotly==5.14.0



python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0

requests==2.28.2
tenacity==8.2.2





